reset to master

+++ http/CurlUtils.cc
@@ -904,7 +904,7 @@ static void process_http_code_helper(long http_code, const string &requested_url
  * @return true if at all worked out, false if it didn't and a retry is reasonable.
  * @throws BESInternalError When something really bad happens.
 */
-static bool eval_http_get_response(CURL *ceh, const string &requested_url, unsigned int &http_code) {
+static bool eval_http_get_response(CURL *ceh, const string &requested_url, long &http_code) {
     BESDEBUG(MODULE, prolog << "Requested URL: " << requested_url << endl);
 
     http_code = 0;
@@ -982,7 +982,7 @@ static void super_easy_perform(CURL *c_handle, int fd) {
     unsigned int attempts = 0;
     useconds_t retry_time = url_retry_time; // 0.25 seconds
     bool success;
-    unsigned int http_code;
+    long http_code;
     do {
         ++attempts;
         BESDEBUG(MODULE,
